fix(scoring): make auto-detect work on the web (Int64List)
Scanning a target and tapping "Finn treff automatisk" found nothing on
the web build — every scan, every program — and silently fell back to
manual marking. The detector built its integral image with an Int64List,
and the web compiler (dart2js) has no 64-bit integer array, so it threw
"Unsupported operation" at runtime. The scanner catches any failure and
returns nothing, so the error never surfaced; it just looked like the
detector found no hits.

Back the integral image with a Float64List instead. The running sums are
whole numbers no larger than 255·width·height, far below 2^53, so a
double holds them exactly — same result on native and web.

This slipped through because `flutter test` runs on the VM and `flutter
build web` only compiles (the throw is at runtime). Add a browser
regression test (`hole_detector_web_test.dart`, `@TestOn('browser')`) and
a CI step that runs it with `flutter test --platform chrome`, and note it
in spec 0040.

+++ lib/features/scoring/domain/hole_detector.dart
@@ -132,7 +132,11 @@ class HoleDetector {
     final w = field.width;
     final h = field.height;
     // Integral image with a zero border: sum over [0,x]×[0,y] at (x+1,y+1).
-    final integral = Int64List((w + 1) * (h + 1));
+    // Float64List, not Int64List: dart2js has no 64-bit integer typed list, so
+    // Int64List throws "Unsupported operation" on the web — which made every
+    // scan fail there. The sums are whole and ≤ 255·W·H, far under 2^53, so a
+    // double holds them exactly.
+    final integral = Float64List((w + 1) * (h + 1));
     for (var y = 0; y < h; y++) {
       var rowSum = 0;
       for (var x = 0; x < w; x++) {
@@ -142,7 +146,7 @@ class HoleDetector {
       }
     }
 
-    int boxSum(int x0, int y0, int x1, int y1) =>
+    double boxSum(int x0, int y0, int x1, int y1) =>
         integral[(y1 + 1) * (w + 1) + (x1 + 1)] -
         integral[y0 * (w + 1) + (x1 + 1)] -
         integral[(y1 + 1) * (w + 1) + x0] +
